BIOS: Properly read sector count for big SATA disks.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@42842 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/PC/BIOS/ahci.c
@@ -668,7 +668,7 @@ static void ahci_port_detect_device(ahci_t __far *ahci, uint8_t u8Port)
 
                 /** @todo update sectors to be a 64 bit number (also lba...). */
                 if (cSectors == 268435455)
-                    cSectors = *(uint16_t *)(abBuffer+(100*2)); // words 100 to 103 (someday)
+                    cSectors = *(uint32_t *)(abBuffer+(100*2)); // words 100 to 103 (someday)
 
                 DBG_AHCI("AHCI: %ld sectors\n", cSectors);
